Fix tests and rename some helpers
