+++ src/helpers/getPetitionAuthor.js
