+++ src/helpers/getPetitionEndDate.js
